fix: Add result path to custom flow
This doesn't fix anything wrong with the flow itself, but I found out
that NOT specifying a result path will overwrite EVERYTHING on the
flow's volitile state. So if any next steps rely on it, which they almost
always do for Gladier, those next steps will suddenly fail.

So if anyone copies this code and wonders why multiple states don't work,
this will fix that.